+++ LiveSegmentation/LiveSegmentation.py
@@ -732,4 +732,4 @@ class LiveSegmentationWidget(ScriptedLoadableModuleWidget):
 class LiveSegmentationTest(ScriptedLoadableModuleTest):
     def runTest(self):
         self.delayDisplay("Live Segmentation module loaded")
-        self.assertEqual(PLUGIN_VERSION, "0.13.0")
+        self.assertEqual(PLUGIN_VERSION, "0.13.1")